fix(package): vue-cropper, move menu jsx
