+++ package.json
@@ -22,7 +22,7 @@
     "viser-vue": "^2.3.3",
     "vue": "^2.5.22",
     "vue-clipboard2": "^0.2.1",
-    "vue-cropper": "^0.4.4",
+    "vue-cropper": "0.4.4",
     "vue-ls": "^3.2.0",
     "vue-router": "^3.0.1",
     "vue-svg-component-runtime": "^1.0.1",
