+++ src/core/use.js
@@ -5,6 +5,7 @@ import config from '@/config/defaultSettings'
 // base library
 import Antd from 'ant-design-vue'
 import Viser from 'viser-vue'
+import VueCropper from 'vue-cropper'
 import 'ant-design-vue/dist/antd.less'
 
 // ext library
@@ -20,3 +21,4 @@ Vue.use(Viser)
 Vue.use(VueStorage, config.storageOptions)
 Vue.use(VueClipboard)
 Vue.use(PermissionHelper)
+Vue.use(VueCropper)
