+++ src/views/account/settings/AvatarModal.vue
@@ -34,12 +34,14 @@
   </a-modal>
 </template>
 <script>
-import { VueCropper } from 'vue-cropper'
+// import { VueCropper } from 'vue-cropper'
 
 export default {
+  /*
   components: {
     VueCropper
   },
+  */
   data () {
     return {
       visible: false,
